loader: Override layer support and settings
Add support for an override layer in the loader. The
usage will become obvious in a few days.

Revamp the logic to detect manifest files in the various
paths that we allow them to be placed.  This code greatly
re-writes loader_get_manifest_files.

Clean up layer functions in loader to follow the proper
coding standard (lower-camel-case for functions).

Also, add the logic to find the layer settings file
using the loader.  That way we can find the proper path
once and provide it to all layers.  This also allows us
to place the settings file in a global directory
so we can override it with the pending layer manager GUI.

Change-Id: I2851dc36e97cd5564cfc57445f4a71b397e8c1a0