layers:Limited POC of synch object tracking
This is the initial version of a single-case update to handle
synchronization tracking for the cmd sequence from GH892. As the code
stands it's only primarily useful for this single cmd sequence and
isn't catching any variety of memory conflicts.

The intention is that if the general idea is sound then I will build on
this initial code to expand/generalize the algorithms used here for all
memory conflict cases.